Restructured database interface and adapters
