+++ src/databaseAdapters/databaseInterfaces/index.ts
@@ -1,0 +1,5 @@
+import IUserDbAdapter from './user';
+
+export default interface IDatabaseAdapter {
+  user: IUserDbAdapter;
+}
