+++ src/sockets/commands/admin/acreateseason.ts
@@ -2,7 +2,7 @@
 
 import { Command } from '../types';
 import { SocketUser } from '../../types';
-import seasonAdapter from '../../../databaseAdapters/season';
+import seasonAdapter from '../../../databaseAdapters/mongoose/season';
 import { sendReplyToCommand } from '../../sockets';
 import { ISeason } from '../../../models/types/season.types';
